requirements.txt
async-timeout==4.0.2
Authlib==1.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Deprecated==1.2.13
Flask==2.0.3
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nsj-gcf-utils==1.0.9
packaging==21.3
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.26.0
six==1.16.0
sqlparams==3.0.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
